+++ src/main/java/com/edu/utadeo/controllers/EmpresaController.java
@@ -3,15 +3,10 @@ package com.edu.utadeo.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.edu.utadeo.modelEntity.Empresa;
@@ -33,19 +28,4 @@ public class EmpresaController {
 	public Empresa detail(@PathVariable long id) {
 		return empresaService.findById(id);
 	}
-	
-	@PostMapping("/")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Empresa save(@RequestBody Empresa e) {
-		return empresaService.save(e);
-	}
-	
-	@DeleteMapping("/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void  delete(@PathVariable long id) {
-		empresaService.delete(id);
-	}
-	
-	@PutMapping("/")
-	
 }
