+++ src/main/java/com/edu/utadeo/modelEntity/Pedido.java
@@ -5,12 +5,9 @@ import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -29,43 +26,4 @@ public class Pedido implements Serializable{
 	@Column
 	@JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
 	private Date fechaEntrega;
-	
-	@ManyToOne(fetch=FetchType.LAZY)
-	@JoinColumn(name="tipo_empleado.id")
-	private Empleado empleado;
-	
-	//getters and setters
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public Date getFechaSolicitud() {
-		return fechaSolicitud;
-	}
-
-	public void setFechaSolicitud(Date fechaSolicitud) {
-		this.fechaSolicitud = fechaSolicitud;
-	}
-
-	public Date getFechaEntrega() {
-		return fechaEntrega;
-	}
-
-	public void setFechaEntrega(Date fechaEntrega) {
-		this.fechaEntrega = fechaEntrega;
-	}
-
-	public Empleado getEmpleado() {
-		return empleado;
-	}
-
-	public void setEmpleado(Empleado empleado) {
-		this.empleado = empleado;
-	}
-	
-	
 }
